requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-cohere
langgraph
requests
python-dotenv
pydantic
cohere
pinecone-client
ollama
spacy
supabase
langchain-supabase
ragas